Refactored code to extract common elements from pattern view models
