+++ BatchRenamer/BatchRenamer.csproj
@@ -121,6 +121,7 @@
     <Compile Include="ViewModels\Patterns\NumberizePatternViewModel.cs" />
     <Compile Include="ViewModels\Patterns\InsertTextPatternViewModel.cs" />
     <Compile Include="ViewModels\Patterns\FindAndReplacePatternViewModel.cs" />
+    <Compile Include="ViewModels\Patterns\PatternViewModelBase.cs" />
     <Compile Include="ViewModels\Patterns\WebMatchPatternViewModel.cs" />
     <Compile Include="Data\Patterns\WebRegexMatch.cs" />
     <Compile Include="ViewModels\ViewModelBase.cs" />
